requirements.txt
pandas
altair
streamlit-option-menu
python-dotenv
dotenv
psycopg2
Unidecode
plotly
geopy